+++ desktop-app/src/config/env.js
@@ -7,10 +7,8 @@ let {
     OPEN_IN_DEFAULT_BROWSER
 } = process.env
 
-// Misc.
-let TELEMETRY_ENABLED = true
-
 // Production mode.
+// Sanity-check.
 if (process.env.NODE_ENV == 'production') {
     DEV_IPFS = false
     DEV_UI = false
@@ -26,7 +24,6 @@ const env = {
     DEV_GATEWAY,
     // Open links in the default browser instead of the embedded browser.
     OPEN_IN_DEFAULT_BROWSER,
-    TELEMETRY_ENABLED
 }