+++ desktop-app/src/index.js
@@ -3,7 +3,7 @@ sourceMap.install();
 import * as path from 'node:path';
 import {spawn, spawnSync} from 'node:child_process';
 
-import {app, BrowserWindow, ipcMain, MessageChannelMain, shell} from 'electron';
+import {app, BrowserWindow, ipcMain, Menu, MessageChannelMain, shell} from 'electron';
 import electronIsDev from 'electron-is-dev';
 import serve from 'electron-serve';
 
@@ -24,8 +24,11 @@ import {
 } from "node:worker_threads"
 
 const { featureFlags, env } = require("./config")
-import telemetry from './telemetry'
 import { ServiceLoggerStream } from './utils';
+
+import telemetry from './telemetry'
+telemetry.configure()
+import './menu'
 
 
 // 
@@ -492,7 +495,6 @@ function fixDappnetDataLocation() {
 fixDappnetDataLocation()
 printBanner()
 parseArguments()
-telemetry.configure()
 configureAutomaticUpdates()
 configureApp()
 serveUI()